requirements.txt
Django==1.9.5
argparse==1.2.1
dj-database-url==0.4.0
gunicorn==19.4.5
psycopg2==2.6.1
wsgiref==0.1.2
Pillow==3.2.0
django-solo==1.1.2
